Fixed PNG
LibPNG expects a proper 2D array, unlike LibJPEG where you can get away with a 'pseudo' 2D array that's just a long 1D array. This requires that you implement a proper iterator class to step through the 2D array, which is tideous and not as elegant as the 1D array. That's what was implemented.